add:some validations in get, delete and put.

+++ src/main/java/com/one/foroapi/domain/service/CommentService.java
@@ -8,7 +8,9 @@ import com.one.foroapi.domain.model.User;
 import com.one.foroapi.domain.repository.CommentRepository;
 import com.one.foroapi.domain.repository.PostRepository;
 import com.one.foroapi.domain.repository.UserRepository;
+import com.one.foroapi.infra.exceptions.notEditableException;
 import com.one.foroapi.util.TimeLimit;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -50,25 +52,32 @@ public class CommentService {
     }
 
     public Comment getCommentById(Long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw new EntityNotFoundException("Comment not found: " + commentId);
+        }
         return commentRepository.findById(commentId).orElse(null);
     }
 
     public void deleteCommentById(Long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw new EntityNotFoundException("Comment not found: " + commentId);
+        }
         commentRepository.deleteById(commentId);
     }
 
     public Comment updateComment(Long commentId, UpdateCommentDTO updateComment) {
         Comment comment = getCommentById(commentId);
-
-        if (!TimeLimit.isEditableWithinTimeLimit(comment.getCreated_at())) {
-            throw new RuntimeException("The topic is not editable after 15 minutes of its creation dat time");
+        if (comment != null) {
+            if (!TimeLimit.isEditableWithinTimeLimit(comment.getCreated_at())) {
+                throw new notEditableException("The comment is not editable after 15 minutes of its creation dat time");
+            }
+            if (updateComment.content() != null) {
+                comment.setContent(updateComment.content());
+            }
+            return commentRepository.save(comment);
+        } else {
+            throw new EntityNotFoundException("Comment not found: " + commentId);
         }
-
-        if (updateComment.content() != null) {
-            comment.setContent(updateComment.content());
-        }
-
-        return commentRepository.save(comment);
     }
 
 }
